@@ -1,34 +1,34 @@
 buildscript {
     repositories {
         jcenter()
+        google()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:1.5.0'
+        classpath 'com.android.tools.build:gradle:4.0.1'
     }
 }
 
 allprojects {
     repositories {
         jcenter()
+        google()
     }
 }
 
 ext {
-    minSdkVersion = 8
-    sampleMinSdkVersion = 15
+    minSdkVersion = 21
+    sampleMinSdkVersion = 21
     targetSdkVersion = 23
-    compileSdkVersion = 23
+    compileSdkVersion = 30
     buildToolsVersion = '23.0.1'
 }
 
 ext.deps = [
         // Android
-        recyclerview         : 'com.android.support:recyclerview-v7:23.0.1',
-        supportv4            : 'com.android.support:support-v4:23.1.0',
-        appcompat            : 'com.android.support:appcompat-v7:23.1.0',
-        cardview             : 'com.android.support:cardview-v7:23.1.0',
+        recyclerview         : 'androidx.recyclerview:recyclerview:1.1.0',
+        appcompat            : 'androidx.appcompat:appcompat:1.2.0',
+        material             : 'com.google.android.material:material:1.2.0',
 
         // External
-        recyclerviewanimators: 'jp.wasabeef:recyclerview-animators:1.3.0',
-        smoothprogressbar    : 'com.github.castorflex.smoothprogressbar:library-circular:1.1.0'
+        recyclerviewanimators: 'jp.wasabeef:recyclerview-animators:4.0.0',
 ]
