@@ -2,16 +2,17 @@ package com.example.paginate;
 
 import android.os.Bundle;
 import android.os.Handler;
-import android.support.v7.widget.GridLayoutManager;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.OrientationHelper;
-import android.support.v7.widget.RecyclerView;
-import android.support.v7.widget.StaggeredGridLayoutManager;
 import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
+
+import androidx.recyclerview.widget.GridLayoutManager;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.OrientationHelper;
+import androidx.recyclerview.widget.RecyclerView;
+import androidx.recyclerview.widget.StaggeredGridLayoutManager;
 
 import com.example.paginate.adapter.RecyclerPersonAdapter;
 import com.example.paginate.data.DataProvider;
@@ -142,7 +143,8 @@ public class RecyclerViewExampleActivity extends BaseActivity implements Paginat
 
             // This is how you can make full span if you are using StaggeredGridLayoutManager
             if (recyclerView.getLayoutManager() instanceof StaggeredGridLayoutManager) {
-                StaggeredGridLayoutManager.LayoutParams params = (StaggeredGridLayoutManager.LayoutParams) vh.itemView.getLayoutParams();
+                StaggeredGridLayoutManager.LayoutParams params =
+                        (StaggeredGridLayoutManager.LayoutParams) vh.itemView.getLayoutParams();
                 params.setFullSpan(true);
             }
         }
