@@ -1,8 +1,12 @@
 package com.paginate.recycler;
 
-/** SpanSizeLookup that will be used to determine the span of loading list item. */
+/**
+ * SpanSizeLookup that will be used to determine the span of loading list item.
+ */
 public interface LoadingListItemSpanLookup {
 
-    /** @return the span of loading list item. */
+    /**
+     * @return the span of loading list item.
+     */
     int getSpanSize();
 }
